requirements.txt
aiogram==3.10.0
playwright==1.45.0
python-dotenv==1.0.1